requirements.txt
# Core Dependencies
python-docx>=0.8.11
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
langchain>=0.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
gpt4all>=0.2.0

# GUI Dependencies
# tkinter  # Usually comes with Python - not installable via pip
ttkthemes>=3.2.0

# Processing Dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0  # For GPU support
transformers>=4.30.0

# Utility Dependencies
tqdm>=4.65.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0

# Testing Dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development Dependencies
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0 